clock v3.1.71: a failed symbol hook latched and never retried
On a freshly started Explorer the mod logged "Got system tray module" and then
"HookSymbols failed", and the clock was never customised again.

GetSystemTrayModuleHandle prefers SystemTray.dll and falls back through
Taskbar.View.dll and ExplorerExtensions.dll to taskbar.dll. On Windows 11 26H2
build 26300 taskbar.dll is always loaded but no longer hosts the winrt::SystemTray
types, and SystemTray.dll is loaded later than Wh_ModInit/Wh_ModAfterInit run. So
the emergency fallback resolved to the wrong module and the hook failed.

That alone would be recoverable - there is a LoadLibraryExW hook that catches the
real module when it arrives. But all three call sites set g_systemTrayModuleHooked
before knowing whether the hook succeeded, and HandleLoadedModuleIfSystemTray is
gated on that flag being clear. One failed attempt against taskbar.dll therefore
locked out the retry permanently.

Now the flag is only latched on success; a failed attempt releases its claim. The
exchange stays, so two concurrent module loads still cannot both hook. Wh_ModInit
also no longer returns FALSE on that failure, since failing to hook a module that
was never the right one is not a reason to refuse to load.

Nothing here is related to the v3.1.70 logging change - that touched only Wh_Log
call sites - but 3.1.70 shipped the settings-block fix without a version bump, so
this bump also guarantees a distinguishable rebuild.

+++ mods/taskbar-ai-quota-fork/taskbar-ai-quota-fork.wh.cpp
@@ -3754,9 +3754,15 @@ static Grid BuildQuotaGrid(QuotaUiInstance& state) {
                         bool haveCursor = GetCursorPos(&cursor) != FALSE;
                         bool trackingHover = wasOpen || refs.reopenToolTipOnMove ||
                                              refs.manualToolTipOpen;
+                        // POINT.x/y are LONG, so std::abs gives a long; the thresholds
+                        // have to be UINT because that is what SystemParametersInfoW
+                        // writes into. Cast at the comparison - they are clamped to
+                        // [2, 32], so this never truncates.
                         if (trackingHover && refs.hasToolTipOpenCursor && haveCursor &&
-                            std::abs(cursor.x - refs.toolTipOpenCursor.x) < toolTipMoveThresholdX &&
-                            std::abs(cursor.y - refs.toolTipOpenCursor.y) < toolTipMoveThresholdY) {
+                            std::abs(cursor.x - refs.toolTipOpenCursor.x) <
+                                (LONG)toolTipMoveThresholdX &&
+                            std::abs(cursor.y - refs.toolTipOpenCursor.y) <
+                                (LONG)toolTipMoveThresholdY) {
                             return;
                         }
                         if (refs.manualToolTipTimer) refs.manualToolTipTimer.Stop();
